% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cement}
\alias{cement}
\title{A 2D cement paste image}
\format{
An integer matrix of 128 rows and 128 columns, with the following
labels for the different segmented phases:
\enumerate{
\item Pores
\item Portlandite
\item Clinker
\item CSH
}
}
\source{
\link{https://doi.org/10.1103/PhysRevE.100.053316}
}
\usage{
cement
}
\description{
A 2D segmented cement paste image, as an integer matrix, which carries the
\code{nmhsa_array} S3 class for more friendly printing and plotting.
}
\keyword{datasets}
