% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmhsa.R
\name{hsa}
\alias{hsa}
\title{Hierarchical simulated annealing}
\usage{
hsa(
  ti,
  dimensions = NULL,
  start_after = NULL,
  stop_after = NULL,
  stop_rejected = NULL,
  stop_at = 1e-06,
  order = NULL,
  distance = NULL,
  cool = 0.9
)
}
\arguments{
\item{ti}{The original training image (2D array).}

\item{dimensions}{Dimensions for the reconstruction (vector of length 2 or
3). Defaults to the dimensions of \code{ti}.}

\item{start_after}{Number of iterations for determining the initial simulated
annealing temperature. Defaults to 0.1\% of the total amount of pixels in
the reconstruction.}

\item{stop_after}{Maximum number of iterations. Defaults to the amount of
pixels in the reconstruction. Can be a vector for different values per
phase.}

\item{stop_rejected}{Maximum number of consecutively rejected iterations.
Defaults to 1\% of the total amount of pixels in the reconstruction.}

\item{stop_at}{Target precision. Defaults to \code{1e-6}.}

\item{order}{Integer vector with the phase order for hierarchical simulation.
Defaults to the least to the most occurring phase.}

\item{distance}{Distance, in pixels, up to which to investigate the
structural descriptors. Defaults to 1/4th of the minimum of \code{dim(ti)} and
\code{dimensions}.}

\item{cool}{Simulated annealing cooling factor. Must be lower than one.
Defaults to \code{0.9}.}
}
\value{
A reconstructed 2D or 3D array.
}
\description{
This is the simplest algorithm available at the moment, which is capable of
multiphase reconstruction using the hierarchical approach (phases are added
one by one).
}
\references{
\link{https://doi.org/10.1103/PhysRevE.100.053316}
}
\seealso{
\code{\link[=mhsa]{mhsa()}}, \code{\link[=nmhsa]{nmhsa()}}
}
