% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmhsa.R
\name{mhsa}
\alias{mhsa}
\title{Multiresolution hierarchical simulated annealing}
\usage{
mhsa(
  ti,
  dimensions = NULL,
  start_after = NULL,
  stop_after = NULL,
  stop_rejected = NULL,
  stop_at = 1e-06,
  order = NULL,
  distance = NULL,
  cool = 0.9,
  levels = 2
)
}
\arguments{
\item{ti}{The original training image (2D array).}

\item{dimensions}{Dimensions for the reconstruction (vector of length 2 or
3). Defaults to the dimensions of \code{ti}.}

\item{start_after}{Number of iterations for determining the initial simulated
annealing temperature. Defaults to 0.1\% of the total amount of pixels in
the reconstruction.}

\item{stop_after}{Maximum number of iterations. Defaults to the amount of
pixels in the reconstruction. Can be a vector for different values per
phase.}

\item{stop_rejected}{Maximum number of consecutively rejected iterations.
Defaults to 1\% of the total amount of pixels in the reconstruction.}

\item{stop_at}{Target precision. Defaults to \code{1e-6}.}

\item{order}{Integer vector with the phase order for hierarchical simulation.
Defaults to the least to the most occurring phase.}

\item{distance}{Distance, in pixels, up to which to investigate the
structural descriptors. Defaults to 1/4th of the minimum of \code{dim(ti)} and
\code{dimensions}.}

\item{cool}{Simulated annealing cooling factor. Must be lower than one.
Defaults to \code{0.9}.}

\item{levels}{Amount of grid levels (or resolutions) to use for the
reconstruction. Defaults to \code{2}. Can be a vector to use different levels
for different phases.}
}
\value{
A reconstructed 2D or 3D array.
}
\description{
This algorithm extends \code{\link[=hsa]{hsa()}} by introducing different grid levels, or
resolutions, which should result in improved and more efficient
reconstruction of larger particles.
}
\references{
\link{https://doi.org/10.1103/PhysRevE.100.053316}
}
\seealso{
\code{\link[=hsa]{hsa()}}, \code{\link[=nmhsa]{nmhsa()}}
}
