% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{phase_merge}
\alias{phase_merge}
\title{Merge neighbouring pixels}
\usage{
phase_merge(img, phase, into)
}
\arguments{
\item{img}{A 2D array.}

\item{phase}{The phase to modify.}

\item{into}{The phase to change pixels to.}
}
\value{
The modified 2D array.
}
\description{
This functionality was originally introduced to remove inclusions of certain
phases within another phase. We are however simply looking for neighbouring
pixels here. As this is still effective with small inclusions, for larger
ones, repeated calls of this function may be required. Note however that this
also affects pixels of the phase to modify that are not inclusions but just
neighbouring pixels of the target phase.
}
\seealso{
\code{\link[=phase_split]{phase_split()}}, \code{\link[=nmhsa]{nmhsa()}}
}
