% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{phase_split}
\alias{phase_split}
\title{Split into subphases}
\usage{
phase_split(img, phase, larger_than, into = max(c(img)) + 1)
}
\arguments{
\item{img}{A 2D array.}

\item{phase}{The phase to modify.}

\item{larger_than}{The threshold for splitting (in pixels).}

\item{into}{The phase to change pixels to.}
}
\value{
The modified 2D array.
}
\description{
This function splits a certain phase into two subphases based on a size
threshold, which makes sense in the reconstruction framework if the two
subphases behave different structurally.
}
\seealso{
\code{\link[=phase_merge]{phase_merge()}}, \code{\link[=nmhsa]{nmhsa()}}
}
