% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapper}
\alias{wrapper}
\title{Access to the Python wrapper examples}
\usage{
wrapper(algorithm)
}
\arguments{
\item{algorithm}{One of "hsa", "mhsa", "nmhsa" or "3d".}
}
\value{
A 2D or 3D array
}
\description{
Access to the Python wrapper examples
}
